% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_subclass}
\alias{add_subclass}
\alias{add_superclass}
\alias{add_class}
\title{Add class}
\usage{
add_subclass(x, class)

add_superclass(x, class)

add_class(x, class)
}
\arguments{
\item{x}{object}

\item{class}{character vector of class names}
}
\value{
object with added class
}
\description{
Extends an object by adding one or more classes to an object
}
\details{
\code{add_subclass} adds classes to the \strong{front} of the class vector

\code{add_superclass} adds classes to the end of the special \code{class} attribure

\code{add_class} is an alias for \code{add_subclass}.
}
