% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{alternate_files}
\alias{alternate_files}
\title{alternate_files are files that have registered extensions but are not the
default.}
\usage{
alternate_files(name)
}
\description{
alternate_files are files that have registered extensions but are not the
default.
}
\seealso{
\code{\link[=conflicts]{conflicts()}}
}
