% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.regex.R
\name{as.regex}
\alias{as.regex}
\alias{as.regex.default}
\alias{as.regex.ext}
\title{Convert to a regular expressions}
\usage{
as.regex(x, ...)

\method{as.regex}{default}(x, ...)

\method{as.regex}{ext}(x, ...)
}
\arguments{
\item{x}{an object to convert to a regular expression}
}
\description{
Converts an object to a regular expression for matching
}
\details{
\code{\link[=as.regex]{as.regex()}} converts \code{pattern}
}
\examples{
  as.regex("pattern")
  as.regex( ext('rds') )

}
