% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_regex.R
\name{as_regex}
\alias{as_regex}
\alias{as_regex.default}
\alias{as_regex.fs_ext}
\title{Convert to a regular expressions}
\usage{
as_regex(x, ...)

\method{as_regex}{default}(x, ...)

\method{as_regex}{fs_ext}(x, ...)
}
\arguments{
\item{x}{an object to convert to a regular expression}
}
\description{
Converts an object to a regular expression for matching
}
\details{
\code{\link[=as_regex]{as_regex()}} converts \code{pattern}
}
\examples{
  as_regex("pattern")
  as_regex( ext('rds') )

}
