% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_get.R
\name{backend_get}
\alias{backend_get}
\alias{backend_get.NULL}
\alias{backend_get.fs_ext}
\alias{backend_get.fs_path}
\alias{backend_get.character}
\title{Get backend}
\usage{
backend_get(x)

\method{backend_get}{`NULL`}(x)

\method{backend_get}{fs_ext}(x)

\method{backend_get}{fs_path}(x)

\method{backend_get}{character}(x)
}
\arguments{
\item{x}{argument used to get the backend with}
}
\value{
\code{backend_get} returns a list for the backend.
}
\description{
Get the cache backend
}
\details{
\code{backend_get} retrieves the definition for the cache backend. The definition
is a list that was created by \code{\link[=cache_register]{cache_register()}}.
}
\seealso{
\itemize{
\item \code{\link[=backend]{backend()}}
\item \code{\link[=as_backend]{as_backend()}}
}
}
