% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_register.R
\name{backend_register}
\alias{backend_register}
\alias{cache_add_extension}
\alias{cache_add_ext}
\title{Register cache backend}
\usage{
backend_register(name, reader = get(paste0("cache_read_", name)),
  writer = get(paste0("cache_write_", name)), ext = name, ...)

cache_add_extension(...)

cache_add_ext(...)
}
\arguments{
\item{name}{string; name for the backend.}

\item{reader}{function for reading from the cache, defaults to
function \code{cache_read_[name]}}

\item{writer}{function; function for writing to the cache,}

\item{ext}{string; the extension used for the file type, usually everything
\strong{after} the final period. Defaults to \code{name}.}

\item{...}{additional name/value pairs to be stored in the backend.}
}
\value{
The value from \code{\link[base:options]{base::options()}} is returned
}
\description{
Register a backend for reading and writing to the cache.
}
\details{
\code{backend_register()} associated an extension with functions for reading and
writing that file type.
\itemize{
\item \code{name} is simply a label by which the backend is referred.
\item \code{reader} is the function used for reading/loading the file into R
\item \code{writer} is the function used for storing the file (to the filesystem)
\item \code{ext} argument should not include the period.
}

It is mostly for writing extensions writing different types of files.
}
\examples{

\dontrun{
  cache_add_ext( 'csv', readr::read_csv, readr::write_csv )
}

}
