% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_set.R
\name{backend_set}
\alias{backend_set}
\title{Set the default backend for cacheing}
\usage{
backend_set(backend)
}
\arguments{
\item{backend}{string; name for the backend.  Note this is not
the file extension but the name used for cacheing.}
}
\description{
Set the default backend for cacheing
}
\details{
\code{backend_set} sets the default method for reading and writing from the
cache. It throws an error if no backend is registered for the methods.
}
\seealso{
\itemize{
\item \code{\link[=backend_register]{backend_register()}}
\item \code{\link[=backend_ls]{backend_ls()}}
}
}
