% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends}
\alias{backends}
\alias{backends_ls}
\alias{backend_ls}
\alias{backends_exts}
\title{Get backends}
\usage{
backends()

backends_ls()

backend_ls()

backends_exts()
}
\value{
\code{backends} returns a list of backends cache backends stored in the
option \code{cache.backends}

character vector of names for the backends

\code{backends_exts} returns a named character vector \code{backend_name} => \code{ext}
}
\description{
Returns (the list of) backends
}
\details{
\code{backends} returns the list of registered backends

\code{backend_ls} lists the names of registered backends.

\code{backends_exts} lists the registered extensions.
}
