% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg.R
\name{bg}
\alias{bg}
\title{Pipeable perform in the background}
\usage{
bg(x, fun, ...)
}
\arguments{
\item{fun}{}

\item{x;}{first}
}
\description{
Performs function according to future::plan()
}
\details{
\code{bg} executed function \code{fun} in the backgrodn
}
\examples{

  plan(multiprocess)
  f <- 5 \%>\% future({Sys.sleep(.)})
  t <- 5 \%>\% bg(Sys.sleep)

}
\seealso{
\itemize{
\item \code{\link[future:future]{future::future()}} and \code{\link[future:plan]{future::plan()}}
}
}
