% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\alias{cache_}
\alias{uncache}
\alias{cache_all}
\alias{uncache_}
\alias{uncache_all}
\title{Cache and uncache an object}
\usage{
cache(object, cache = getOption("cache", "cache"),
  timestamp = getOption("timestamp"), envir = parent.frame())

cache_(name, cache = getOption("cache", "cache"),
  timestamp = getOption("timestamp"), envir = parent.frame())

uncache(name, cache = getOption("cache", "cache"),
  timestamp = getOption("timestamp"), envir = parent.frame(),
  overwrite = TRUE)

cache_all(...)

uncache_(...)

uncache_all(cache = getOption("cache", "cache"), envir = parent.frame())
}
\arguments{
\item{object}{object to cache}

\item{cache}{character; path to cache directory. The default is
\code{getOption('cache')} or else simply \code{cache}}

\item{timestamp}{string or function (of 0 arguments); timestamp to add to
the filename.}

\item{envir}{where to uncache the object. Defaults to the caller's environment}

\item{name}{string; name for the object}

\item{overwrite}{logical; whether to overwrite if it \code{name} already exists
on the environment.}

\item{...}{additional arguments

\code{cache}/\code{cache_} and \code{uncache}/\code{uncache_} save and restore single objects to
the a \code{cache} directory with \code{saveRDS}.

The \code{cache} defaults to the global option 'cache' otherwise
the \code{cache} in the working directory is used. This follows the behavior
in the \emph{ProjectTemplate} package.

\code{uncache_all} restores all files in a cache directory.

\code{cache} also allows for timestamps. The default is to use the global
option \code{timestamp}. This can either be a character vector or a function
of zero arguments that returns an unary length character vector. (Only the
first value is used.)  Common practice is to use \code{Sys.Date} or
\code{Sys.time} for creating the timestamp.}
}
\description{
Caches or uncaches and object in a cache directory
}
\examples{
  \dontrun{
     cache( myfile )                      # cache/myfile.rds
     cache( myfile, "mycache" )           # mycache/myfile.rds
     cache( myfile, "mycache", Sys.Date ) # mycache/myfile-YYYY-MM-DD.rds

     # EXPLICIT USE OF timestamp
     options( timestamp = Sys.Date )
     cache( myfile, "mycache" )          # mycache/myfile-YYYY-MM-DD.rds

     uncache(myfile)
     uncache_("myfile")
  }

}
\seealso{
\code{base::saveRDS} \cr
}
