% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_all.R
\name{cache_all}
\alias{cache_all}
\title{cache_all}
\usage{
cache_all(..., envir = parent.frame())
}
\arguments{
\item{...}{additional arguments passed to \code{\link[=cache]{cache()}}}

\item{envir}{environment from which to find objects to save.}
}
\description{
cache all objects
}
\seealso{
\itemize{
\item \code{\link[base:save]{base::save()}}
}
}
