% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_backend.R
\name{cache_backend}
\alias{cache_backend}
\title{Get/Set the default cache backend}
\usage{
cache_backend(backend = NULL)
}
\arguments{
\item{backend}{string; name for the backend.  Note this is not
the file extension but the name of a registered backend.}
}
\description{
Get/Set the default cache backend
}
\details{
\code{cache_backend()} gets or sets the default method(s) for reading and writing
from/to the cache. Without an argument, it returns the current default
backend. With a string \code{backend} argument, it sets the default methods to
those associated with that backend

\code{cache_backend()} throws an error if no backend is registered for the
\code{backend} argument. Register that backend with \code{\link[=cache_register]{cache_register()}} first.
}
\seealso{
\itemize{
\item \code{\link[=cache_register]{cache_register()}}
\item \code{\link[=backends_ls]{backends_ls()}}
}
}
