% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_backend_register.R
\name{cache_backend_register}
\alias{cache_backend_register}
\alias{cache_add_extension}
\alias{cache_add_ext}
\title{Register cache backend}
\usage{
cache_backend_register(name, reader = get(paste0("cache_read_", name)),
  writer = get(paste0("cache_write_", name)), ext = name)

cache_add_extension(...)

cache_add_ext(...)
}
\arguments{
\item{name}{string; name for the backend.}

\item{ext}{string; the extension used for the file type, usually everything
\strong{after} the final period. Defaults to \code{name}.}

\item{read}{function for reading from the cache, defaults to
function \code{cache_read_[extension]}}

\item{write}{function; function for writing to the cache,}
}
\value{
The value from \code{\link[base:options]{base::options()}} is returned
}
\description{
Register a backend for reading and writing to the cache.
}
\details{
\code{cache_backend_register()} registers a extension with functions for reading and
writing that file type. The \code{ext} argument should not include the period.

It is mostly for writing extensions writing different types of files.
}
\examples{

\dontrun{
  cache_add_ext( 'csv', readr::read_csv, readr::write_csv )
}

}
