% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_create.R
\name{cache_create}
\alias{cache_create}
\title{cache_create}
\usage{
cache_create(path = here::here() \%>\% fs::path(cache_name()))
}
\arguments{
\item{path}{path to create cache directory}
}
\description{
Create a directory for files
}
\details{
Create a cache directory at \code{path}. If the path already exists, a warning is
provided.
}
