% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_delete.R
\name{cache_delete}
\alias{cache_delete}
\title{Delete cache directory}
\usage{
cache_delete(path = cache_path(), warn = FALSE)
}
\arguments{
\item{path}{cache directory path; defaults to \code{\link[=cache_path]{cache_path()}}}

\item{warn}{logical; enable warnings}
}
\description{
Delete cache directory including all persisted files and metadata
}
\details{
\code{cache_delete} unceremoniously deletes the cache including all persisted
files and metadata.
}
\seealso{
\itemize{
\item \code{\link[=cache_rm]{cache_rm()}} : for removing individual objects from cache
\item \code{\link[=cache_rm_all]{cache_rm_all()}} : for removing all objects from cache but keeping the directory
\item \code{\link[fs:dir_delete]{fs::dir_delete()}} : which is used by \code{cache_delete}
}
}
