% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_dir.R
\name{cache_dir}
\alias{cache_dir}
\title{cache_dir}
\usage{
cache_dir(cache = getOption("cache", "cache"), ...)
}
\arguments{
\item{cache}{string; cache directory}
}
\description{
List contents of the cache directory
}
\details{
Similar to \code{ls} but defaults to viewing the cache directory.
}
\seealso{
\code{\link[=cache]{cache()}}, \code{\link[=uncache]{uncache()}}
}
