% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ext.R
\name{cache_ext}
\alias{cache_ext}
\title{extension for cache files}
\usage{
cache_ext()
}
\value{
An \code{\link[=ext]{ext()}} object for the current backend.
}
\description{
The filename extension used for the current backend.
}
\details{
\code{cache_ext()} returns the extensions associated with the default cache
backend. The difference between a \emph{fs} and \emph{cache} extension is that the
\emph{cache} extension may contain the \code{.}.
}
\examples{
  cache_ext()

}
\seealso{
\itemize{
\item \code{\link[=cache_backend]{cache_backend()}}
}
}
