% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_find.R
\name{cache_find}
\alias{cache_find}
\title{cache_find}
\usage{
cache_find(path = ".")
}
\description{
Finds cache by searching up the directory structure
}
\seealso{
\code{\link[rprojroot:find_root]{rprojroot::find_root()}}
}
