% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ls.R
\name{cache_ls}
\alias{cache_ls}
\alias{cache_list}
\alias{print.cache_list}
\alias{cache_dir}
\alias{cache_info}
\title{cache_ls}
\usage{
cache_ls(cache = cache_path(), ...)

cache_list(x)

\method{print}{cache_list}(x, ...)

cache_dir(path = cache_path(), ...)

cache_info(cache = getOption("cache", "cache"), ...)
}
\arguments{
\item{cache}{string; cache directory (Default: \code{\link[=cache_path]{cache_path()}} )}

\item{...}{additional arguments passed to \code{\link[fs:dir_ls]{fs::dir_ls()}}}
}
\value{
character vector (names of objects)
}
\description{
List cached files
}
\details{
Similar to \code{ls} but lists the objects in the cache directory that can be
loaded.

\code{cache_list} creates a cache_list object for which a print method can exist.

\code{\link[=cache_dir]{cache_dir()}} lists all files in the cache, supported or not.
}
\seealso{
\itemize{
\item \code{\link[=unsupported_paths]{unsupported_paths()}} and \code{\link[=unsupported_files]{unsupported_files()}}
\item \code{\link[=cache]{cache()}}, \code{\link[=uncache]{uncache()}}
\item \code{\link[fs:dir_ls]{fs::dir_ls()}}
}
}
