% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_set_name.R
\name{cache_name}
\alias{cache_name}
\alias{cache_get_name}
\alias{cache_set_name}
\title{cache_name}
\usage{
cache_name()

cache_get_name()

cache_set_name(name)
}
\arguments{
\item{name}{string; name for the cache}
}
\description{
Get or set the default \emph{name} for the cache directory (not the path). The
default is 'cache'.
}
\details{
\code{cache_set_name} can be used provide an alternative name other than \code{cache}.
Some useful names are '.cache' (hidden) or '.rcache' (possibly avoiding conflicts).

\code{\link[=cache_name]{cache_name()}} is an alieas for \code{\link[=cache_get_name]{cache_get_name()}} or \code{cache_name} return this value.
}
\seealso{
\itemize{
\item \code{\link[=cache_path]{cache_path()}}
}
}
