% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_path}
\alias{cache_path}
\title{cache_path}
\usage{
cache_path(x = NULL)
}
\arguments{
\item{x}{string; path of the cache directory}
}
\value{
string; path to the cache directory
}
\description{
Get or set the cache path
}
\details{
If \code{x} is \code{NULL}, returns the \code{getOption('cache','cache')}. Otherwise, the
cache path is set to \code{x}.

While \code{getOption('cache','cache')} can be used, this provides an interface
that can allow the cache path information to change in one place
}
\seealso{
\itemize{
\item \code{\link[=cache]{cache()}}, \code{\link[=uncache]{uncache()}}
}
}
