% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_get_reader.R, R/cache_path_ext.R
\name{cache_get_reader}
\alias{cache_get_reader}
\alias{cache_get_writer}
\alias{cache_path_ext}
\title{cache_get_reader, cache_get_writer}
\usage{
cache_get_reader(path)

cache_get_writer(path)

cache_path_ext(path)

cache_path_ext(path)
}
\arguments{
\item{path}{string; path to file.}

\item{path}{string; path to file}
}
\description{
Get the read or write function for a path.

Get the \emph{registered} cache extension string from the path.
}
\details{
\link{cache_get_reader{}} and \code{\link[=cache_get_writer]{cache_get_writer()}} return the function for reading
and writing based on a given path.

\code{\link[=cache_path_ext]{cache_path_ext()}} gets the entire extension from the cache path, either a
full path or simply a filename. The extension has to be known and registered
using \code{\link[=cache_add_ext]{cache_add_ext()}}.
}
\seealso{
\itemize{
\item \code{\link[=cache_path_ext]{cache_path_ext()}}
}

\itemize{
\item \code{\link[=cache_add_ext]{cache_add_ext()}}
\item \code{\link[fs:path_ext]{fs::path_ext()}}
}
}
