% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_register.R
\name{cache_register}
\alias{cache_register}
\alias{cache_backend_register}
\title{Register cache backend}
\usage{
cache_register(name, reader = get(paste0("cache_read_", name)),
  writer = get(paste0("cache_write_", name)), ext = name, ...)

cache_backend_register(name, reader = get(paste0("cache_read_", name)),
  writer = get(paste0("cache_write_", name)), ext = name, ...)
}
\arguments{
\item{name}{string; name for the backend.}

\item{reader}{function for reading from the cache, defaults to
function \code{cache_read_[name]}}

\item{writer}{function; function for writing to the cache,}

\item{ext}{string; the extension used for the file type, usually everything
\strong{after} the final period. Defaults to \code{name}.}

\item{...}{additional name/value pairs to be stored in the backend.}
}
\value{
The value from \code{\link[base:options]{base::options()}} is returned
}
\description{
Register a backend for reading and writing to the cache.
}
\details{
\code{cache_register()} defines a backend for storing and retrieving data using
cache.  Ultimately, it associated an filename extension with functions for
reading and writing cache files with that extension.

Unless you are writing a new backend, you will not need to use this function.

\code{cache_backend_register} is an alias for \code{cache_register}.
}
\examples{

  \dontrun{
    cache_register( 'csv', readr::read_csv, readr::write_csv )
  }

}
