% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_rm.R
\name{cache_rm}
\alias{cache_rm}
\alias{cache_rm_all}
\title{cache_rm}
\usage{
cache_rm(..., list = character())

cache_rm_all()
}
\arguments{
\item{...}{objects to be removed as names (unqoted) or character strings
(quoted).}

\item{list}{character vector naming objects to be removed.}
}
\description{
Remove item(s) from the cache
}
\details{
\code{\link[=cache_rm]{cache_rm()}} removes objects from the cache.  Objects can be expressed as
character strings, or in the argument \code{list} or through a combination of
both.
}
\examples{

\dontrun{

  cache(iris)
  cache_rm(iris)
  cache(iris)
  cache_rm('iris')

  cache(iris);cache(mtcars)
  cache_ls()
  cache_rm(iris,mtcars)
  cache_ls()


  cache(iris);cache(mtcars)
  cache_rm( c('iris','mtcars') )
  cache_ls()
}

}
