% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_set}
\alias{cache_set}
\alias{cache_get}
\alias{cache_path}
\title{Set or Get Cache Directory}
\usage{
cache_set(path)

cache_get()

cache_path(path = NULL)
}
\arguments{
\item{path}{string; path to the cache directory}
}
\value{
string; path to the cache directory. `set_path`` does this invisibly
}
\description{
Set or gets the cache directory
}
\details{
\code{cache_set} sets the cache directory, creating it if it doesn't exist.

'cache_get' returns the cache directory.

\code{cache_path} will get or set the cache path depending on whether \code{path} is
provided.
}
\seealso{
\itemize{
\item \code{\link[=cache]{cache()}}, \code{\link[=uncache]{uncache()}}
}
}
