% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_set_name.R
\name{cache_set_name}
\alias{cache_set_name}
\alias{cache_get_name}
\alias{cache_name}
\title{cache_set_name, cache_get_name}
\usage{
cache_set_name(name)

cache_get_name()

cache_name()
}
\arguments{
\item{name}{string; name for the cache}
}
\description{
Get and set the default name for the cache directory (not the path)
}
\details{
\code{cache_set_name} can be used provide an alternative name other than \code{cache}.
Some examples are '.cache' (hidden) or '.rcache' (possibly avoiding conflicts).

\code{cache_get_name} or \code{cache_name} return this value.
}
\seealso{
\itemize{
\item \code{\link[=cache_set]{cache_set()}}, \code{\link[=cache_get]{cache_get()}}, \code{\link[=cache_path]{cache_path()}}
}
}
