% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_use_rds.R
\name{cache_use_rds}
\alias{cache_use_rds}
\alias{cache_write_rds}
\alias{cache_read_rds}
\title{Cache object using RDS Files}
\usage{
cache_use_rds()

cache_write_rds(object, name = deparse(substitute(object)),
  cache = cache_find(), ...)

cache_read_rds(name, cache = find_cache(), ...)
}
\arguments{
\item{name}{string; name of object.}

\item{cache}{path to the cache directory}

\item{...}{additional arguments passed to \code{\link[=save_rds]{save_rds()}} / \code{\link[base:saveRDS]{base::saveRDS()}}}

\item{object;}{object to cache}
}
\description{
Cache objects using RDS \code{\link[base:saveRDS]{base::saveRDS()}}
}
\details{
This is the default method for caching objects.

\code{cache_use_rds} sets RDS as the default caching mechanism to use RDS.

The \code{cache_write_rds} writes objects to the cache using RDS
}
\examples{

\dontrun{
  cache_use_rds
}


dontrun{
   cache_use_rds()
   cache(mtcars)
   if( exists('mtcars') ) rm(mtcars)
   uncache(mtcars)
   fs::dir_ls( cache_find() )
}

}
