% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_open.R
\name{cache_view}
\alias{cache_view}
\title{Open Cache Directory}
\usage{
cache_view(cache = cache_path(), browser = getOption("browser"))
}
\arguments{
\item{cache}{string or path to cache directory. (Default: \code{\link[=cache_path]{cache_path()}})}

\item{browser}{string; name of program to be used as the HTML browser.  See
\link[fs:file_show]{fs::file_show}}
}
\description{
Opens the cache directory
}
\seealso{
\itemize{
\item \code{\link[fs:file_show]{fs::file_show()}} which this function uses
}
}
