% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached.R
\name{cached}
\alias{cached}
\alias{cached.default}
\alias{print.cached}
\title{cached}
\usage{
cached(x, ...)

\method{cached}{default}(x, ...)

\method{print}{cached}(x, ...)
}
\arguments{
\item{x}{object to subclass}
}
\description{
subclass for working with cached objects: path, dir, file, name, ext, etc.
}
\details{
These are essentially used for providing special methods for cache objects vs
their \strong{fs} counterparts.

The biggest difference is \emph{cached} ext vs a fs extension
}
