% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_ext.R
\name{cached_ext}
\alias{cached_ext}
\alias{print.cached_ext}
\alias{as_cached_ext}
\alias{as_cached_ext.character}
\alias{as_cached_ext.cached_path}
\alias{as_cached_ext.cached_file}
\title{cached_ext - cached objects file extension}
\usage{
cached_ext(x = character())

\method{print}{cached_ext}(x, ...)

as_cached_ext(x, exts)

\method{as_cached_ext}{character}(x, exts = NULL)

\method{as_cached_ext}{cached_path}(x, exts = backends_exts())

\method{as_cached_ext}{cached_file}(x, exts = backends_exts())
}
\arguments{
\item{x}{object to make a cached_ext object from usually a string or character}

\item{...}{additional arguments}

\item{exts}{character; vector of allowed extensions.}
}
\description{
A character object representing the extension of a cached object.
}
\details{
The \strong{cached_extension} of an object is a file extension if it is one of the
registered file extensions.  For a generic extension object, see \code{\link[=ext]{ext()}}.

The extension of a cached object is not limited to everything following the
final period in an cached object. (This is usually the case.)

A \code{cached_ext} is character vector representing one or more extensions of
cached objects

\code{as_cached_ext} tries to converts unknown types of \code{x} to \code{cached_path}
before coercing to a \code{cached_ext}.

\code{as_cached_ext} returns the \strong{registered} backend extensions for \code{x} for
\code{cached_path} and \code{cached_file} arguments.
}
\examples{
  # cached_ext()
  cached_ext("rds")
  cached_ext(1)

  as_cached_ext('iris.rds')
  as_cached_ext('cache/iris.rds')

  as_cached_ext( cached_file('iris.rds') )
  as_cached_ext( cached_path('cache/iris.rds') )

  path <- cached_path( c("cache/iris.rds", "cached/mtcars.fst") )
  as_cached_ext(path)

  files <- cached_path( c("iris.rds", "mtcars.fst") )
  as_cached_ext(files)

}
