% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_file.R
\name{cached_file}
\alias{cached_file}
\title{Create a cached_file}
\usage{
cached_file(x, ext = backend_ext())
}
\arguments{
\item{x}{character}
}
\value{
a \code{fs::fs_path} object of the filename. The extension
}
\description{
Create a name for an object.
}
\details{
A cached_file is an \code{fs_path} object for the object. It is the filename
without the path information.
}
\examples{
 as_cached_file(iris)   # e.g. "iris.rds"
 as_cached_file('iris') # e.g. "iris.rds"

 cached_file('iris')    # "fs_path"   "character"

}
