% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_name.R
\name{cached_name}
\alias{cached_name}
\alias{print.cached_name}
\alias{as_cached_name}
\alias{as_cached_name.character}
\alias{as_cached_name.fs_path}
\title{Name for cached object}
\usage{
cached_name(x)

\method{print}{cached_name}(x, ...)

as_cached_name(x = NULL)

\method{as_cached_name}{character}(x)

\method{as_cached_name}{fs_path}(x)
}
\arguments{
\item{...}{additional arguments}
}
\value{
a \code{cached_name} object
}
\description{
User-friendly object name. It doesn't contain either path or ext information
}
\details{
\code{as_cached_name.fs_path} handles \code{cached_file} and \code{cached_path} objects
}
\examples{

  cached_name('iris')

  as_cached_name('iris')
  as_cached_name(iris)

  as_cached_file("iris.rds") \%>\% as_cached_name()

  cached_file("iris.rds") \%>\% as_cached_name()
  cached_path("iris.2.rds") \%>\% as_cached_name()

}
\seealso{
\itemize{
\item \code{\link[=cached_file]{cached_file()}}
\item \code{\link[=cached_path]{cached_path()}}
\item \code{\link[=cached_ext]{cached_ext()}}
}
}
