% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache}
\usage{
clear_cache(path = ".cache")
}
\arguments{
\item{path}{(Character scalar)
Directory to clear; defaults to the cache package's default directory.}
}
\description{
Clears a file-based memory cache.
}
