% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{collapse_comma}
\title{Collapse \emph{vector} to a \emph{string}.}
\usage{
collapse(x, sep = "")

collapse_comma(x)
}
\arguments{
\item{x}{R objects to be collapsed, converting to character vectors as
necessary.}

\item{sep}{an optional character string to separate the results.}
}
\description{
Collapses a vector (coercing to character if necessary) into a string.
}
\details{
This is implemented as
\code{ Reduce( function(x,y) paste(x,y, sep=sep), x ) }

and is different from paste as this operates on a single vector rather than
a list of vectors.

\code{collapse_comma()} uses a sep argument of '\code{, }' (commma and a space). This
is mostly useful for pretty printing and code legibility.
}
\examples{

 collapse( letters )
 collapse( letters, sep=" ")
 collapse( letters, ", " )
 collapse_comma( letters )
 
 \dontrun{ 
   collapse( 'a', 'b', 'c' )  # error
 }
 
}
\seealso{
\code{\link[base:paste]{base::paste()}} or \code{\link[base:paste0]{base::paste0()}} with argumetn \code{collapse=''}
\code{\link[base]{paste0}} with \code{collapse=""}
(Why is this different than paste?)
}
\keyword{manip}
