% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_file.R
\name{default_file}
\alias{default_file}
\alias{default_path}
\title{Defautl cached name and file}
\usage{
default_file(name)

default_path(name)
}
\arguments{
\item{name}{string; object name}
}
\description{
Report default file and path given a name
}
\details{
Provides the default name and path for default cache backend.
}
