% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{ext_to_backend}
\alias{ext_to_backend}
\title{Get backend name(s) from extension(s)}
\usage{
ext_to_backend(ext)
}
\arguments{
\item{ext}{string}
}
\description{
Get backend name(s) from extension(s)
}
\examples{

 ext_to_backend("rds")  # rds
 ext_to_backend("none") # NULL

 ext_to_backend( c('rds','sodium.rds') )

}
