% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_regex.R
\name{ext_to_regex}
\alias{ext_to_regex}
\title{Convert ext to regex}
\usage{
ext_to_regex(ext)
}
\arguments{
\item{ext}{string}
}
\description{
Converts an ext to regex; an extension to \code{\link[stringr:regex]{stringr::regex()}}
}
\details{
Creates a regular expression out of an extension. This will match the end
of a file. It has to be a regex rather than fixed in order to match to the
end of the filename.
}
\examples{

ext_to_regex("rds")     # "\\\\.rds$"
ext_to_regex( c("rds", "fst") )
}
\seealso{
\itemize{
\item \code{\link[stringr:regex]{stringr::regex()}}
}
}
