% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{has_conflict}
\alias{has_conflict}
\title{Determine if there is any conflicting files}
\usage{
has_conflict(name, ext = backend_ext())
}
\arguments{
\item{name}{string}
}
\description{
Determine if there is any conflicting files
}
