% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cache_root.R
\docType{data}
\name{is_cache_root}
\alias{is_cache_root}
\title{is_cache_root}
\format{
An object of class \code{root_criterion} of length 5.
}
\usage{
is_cache_root
}
\description{
rprojroot criterion for finding cache directory.
}
\keyword{datasets}
