% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_cache}
\alias{load_cache}
\title{Load memory cache}
\usage{
load_cache(file)
}
\arguments{
\item{file}{(Character scalar) File path.}
}
\description{
Loads a memory cache from file,
producing an object of class \code{cache_memory}.
}
