% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRDS.R
\name{loadRDS}
\alias{loadRDS}
\alias{loadRDSdir}
\title{loadRDS}
\usage{
loadRDS(file, envir = parent.frame(), verbose = FALSE, ...)

loadRDSdir(dir, pattern = ".rds$", envir = parent.frame(),
  recursive = FALSE, ..., verbose = FALSE)
}
\arguments{
\item{file}{character; a path to a file}

\item{envir}{environment; the environment where the data should be loaded}

\item{verbose}{logical; whether to emit detailed output}

\item{...}{additional arguments passed to \code{\link{readRDS}}}

\item{dir}{character; a path to a directory}

\item{pattern}{character; patterns for matching RDS files, default ".rds$"}

\item{recursive}{logical; whehter to recurse into sub-directories.

\code{loadRDS} loads an individual {RDS} file; \code{loadRDSdir} loads all
the rds files from a directory.

A limitation with readRDS as opposed with loadRDS is that it requires a
name assignment. Very often that it is inconvenient. This assigns the
filename to the basename of the file with the extension removed.}
}
\description{
\code{load}-like functionality for RDS files.
}
\examples{
 # -th

}
\seealso{
\code{\link[base]{readRDS}} \cr
\code{\link{file_basename}}
}
