% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_path}
\alias{manifest_path}
\alias{manifest_exists}
\alias{manifest_create}
\alias{manifest_read}
\alias{manifest_write}
\alias{manifest_get}
\alias{manifest_rm}
\alias{manifest_rebuild}
\title{cache manifest}
\usage{
manifest_path()

manifest_exists()

manifest_create(overwrite = FALSE)

manifest_read()

manifest_write(object)

manifest_get()

manifest_rm(name)

manifest_rebuild()
}
\description{
An account of cached objects
}
\details{
The manifest is the record of cached objects. It is simply a list. Each
element is a list of attributes associated with the object and is given the
same name as the object. The primary function of the manifest is to maintain
a mapping of object names to filenames and is checked to maintain uniqueness.

Whenever objects are added or removed from the cache, the MANIFEST is updated.
It should always match the objects saved in the cache directory.

Internally, the manifest is stored as a YAML file and is located at
\code{meta_path()\MANIFEST} usually \code{cache/.meta/MANIFEST}.
}
\seealso{
\itemize{
\item \link{meta}
}
}
