% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_find}
\alias{meta_find}
\alias{meta_path}
\alias{meta_exists}
\title{cache meta directory}
\usage{
meta_find(cache = cache_path())

meta_path(cache = cache_path())

meta_exists()
}
\description{
The cache meta directory hold information about objects stoded in the cache
}
\details{
\code{meta} is a directory for holding supplementary information
}
