% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{name_to_reader}
\alias{name_to_reader}
\title{#' Get backend name(s) from extension(s)
#'
#' @param ext string
#'
#' @examples
#'
#'  ext_to_backend("rds")  # 'rds'
#'  ext_to_backend("none") # NULL
#'
#'  ext_to_backend( c('rds','sodium.rds') )
#'
#' @export}
\usage{
name_to_reader(name)
}
\description{
#' Get backend name(s) from extension(s)
#'
#' @param ext string
#'
#' @examples
#'
#'  ext_to_backend("rds")  # 'rds'
#'  ext_to_backend("none") # NULL
#'
#'  ext_to_backend( c('rds','sodium.rds') )
#'
#' @export
}
