% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parenthesize.R
\name{parenthesize}
\alias{parenthesize}
\alias{bracketize}
\alias{bracize}
\title{parenthesize (Taken from base.tools)}
\usage{
parenthesize(x)

bracketize(x)

bracize(x)
}
\arguments{
\item{x}{arguments to encase in parentheses}
}
\value{
character
}
\description{
encase in parenthseis, brackets, or braces
}
\details{
\code{x} is coerced to character and encased in parens, brackets or braces
}
\references{
\url{http://english.stackexchange.com/questions/3379/bracket-vs-brace}
}
