% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_name.R
\name{path_to_ext}
\alias{path_to_ext}
\title{path_to_ext}
\usage{
path_to_ext(path = fs::dir_ls(cache_path()), exts = backend_exts())
}
\arguments{
\item{path}{path filepath}
}
\description{
path_to_ext
}
\details{
path_to_ext
}
