% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_name.R
\name{path_to_name}
\alias{path_to_name}
\title{path_to_name}
\usage{
path_to_name(paths = fs::dir_ls(cache_path()))
}
\arguments{
\item{paths}{character; fs paths to convert to object names}
}
\value{
named character vector; filename => object name
}
\description{
Convert path(s) to name(s)
}
\details{
Finds the (object) names associated with a path. The name is discovered by
}
\note{
\code{path_to_name} is not exported
}
