% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save_cache}
\alias{save_cache}
\title{Save memory cache}
\usage{
save_cache(x, file)
}
\arguments{
\item{x}{Object of class \code{cache_memory}.}

\item{file}{(Character scalar) Output file path.}
}
\description{
Saves a memory cache from file.
}
