% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{(Pipeable) set single options}
\usage{
set_option(...)
}
\value{
The options value, invisibly
}
\description{
(Pipeable) set single options
}
\examples{

 x <- set_option( cache_key = "my key" )
 x
 getOption("cache_key")
}
