% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-squote.R
\name{squote}
\alias{squote}
\title{squote ((Taken from base.tools)}
\usage{
squote(x)
}
\arguments{
\item{x}{character; values to enquote

These are the same as \code{base::sQuote} and
except fancy quotes are not used.

Fancy quotes are temporarily disabled for this function.}
}
\description{
single or double quote text without using Fancy Quotes
}
\seealso{
\code{\link[base]{sQuote}} \cr
}
