% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupported_paths.R
\name{unsupported_paths}
\alias{unsupported_paths}
\alias{unsupported_files}
\title{unsupported_paths}
\usage{
unsupported_paths(path = cache_path())

unsupported_files(...)
}
\arguments{
\item{path}{character or \code{\link[fs:path]{fs::path()}} pointing to a directory or file.
(Default: \code{\link[=cache_path]{cache_path()}})}
}
\value{
character vector of unsupported paths/files.
}
\description{
list unsupported_files, i.e. those without a registered backend.
}
\details{
\code{unsuppported_paths()} checks to see whether the paths given by \code{path} have
registered extensions and can therefore be written or read.

We cannot get \code{unsupported_exts()} since we can't, in general, discern
where the filename ends and the extension begins.
}
\examples{


}
\seealso{
\itemize{
\item \code{\link[=cache_path]{cache_path()}}
}
}
