% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cache.R
\name{use_cache}
\alias{use_cache}
\title{use_cache}
\usage{
use_cache(gitignore = ignore, rbuildignore = ignore, ignore = TRUE)
}
\arguments{
\item{gitignore}{logical; whether to put `.cache`}

\item{rbuildignore}{}

\item{path}{string; directory in which to place the cache}
}
\description{
use_cache
}
