% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{get_backends}
\alias{get_backends}
\alias{backend_get}
\alias{backend_ls}
\alias{backend_exts}
\title{Backends}
\usage{
get_backends()

backend_get(backend)

backend_ls()

backend_exts()
}
\arguments{
\item{backend}{string; name for backend}
}
\value{
\code{get_backends} returns a list of backends cache backends stored in the
option \code{cache.backends}

\code{backend_get} returns a list for the backend.

character vector of names for the backends

\code{backend_exts} returns a named character vector \code{backend_name} => \code{ext}
}
\description{
Functions to working with cache backends
}
\details{
A backend for a cache contains complete details about storing and
retrieving files

\code{get_backends} returns the list of registered backends

\code{backend_get} retrieves the definition for the cache backend. The definition
is a list that was created by \code{\link[=backend_register]{backend_register()}}.

\code{backend_ls} lists the names of registered backends.

\code{backend_exts} lists the registered extensions.
}
\seealso{
\code{\link[cache:backend_register]{cache::backend_register()}}
}
