% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backends}
\alias{backends}
\alias{backend_get}
\alias{backend_ls}
\alias{backend_exts}
\title{Various functions for working with cache backends.}
\usage{
backends()

backend_get(backend = cache_backend())

backend_ls()

backend_exts()
}
\arguments{
\item{backend}{string; name for backend}
}
\value{
\code{backends} returns a list of backends cache backends stored in the
option \code{cache.backends}

\code{backend_get} returns a list for the backend.

character vector of names for the backends

\code{backend_exts} returns a named character vector \code{backend_name} => \code{ext}
}
\description{
Various functions for working with cache backends.
}
\details{
A backend for a cache contains complete details about storing and
retrieving files

\code{backends} returns the list of registered backends

\code{backend_get} retrieves the definition for the cache backend. The definition
is a list that was created by \code{\link[=cache_register]{cache_register()}}.

\code{backend_ls} lists the names of registered backends.

\code{backend_exts} lists the registered extensions.
}
