% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_check.R
\name{cache_check}
\alias{cache_check}
\title{cache_check}
\usage{
cache_check(cache = cache_path())
}
\arguments{
\item{cache}{}
}
\description{
Check the cache for problems
}
\details{
\code{cache_check} checks the cache for potential problems.
}
