% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_create.R
\name{cache_create}
\alias{cache_create}
\title{cache_create}
\usage{
cache_create(path)
}
\arguments{
\item{path}{path or sting to create cache directory}
}
\description{
Create a directory for files
}
\details{
Create a cache directory at \code{path}. If the path already exists, a warning is
provided. If path does not exist it is created.
}
\note{
On Windows 7 & 10, it is possible to create a directory/folder \code{.meta}
in File Explorer by typing the name as \code{.meta.} (ie. with a trailing period).
\itemize{
\item \href{https://superuser.com/a/1156507/479495}{SO:create-rename-a-file-folder-that-begins-with-a-dot-in-windows}
}
}
