% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ext.R
\name{cache_ext}
\alias{cache_ext}
\alias{cache_ext.NULL}
\title{Default cache extensions}
\usage{
cache_ext(x)

\method{cache_ext}{`NULL`}(x)
}
\value{
An \code{\link[=ext]{ext()}} object for the current backend.
}
\description{
Get the file/path extension used for the current/defaul cache backend.
}
\details{
\code{cache_ext()} returns the extensions associated with the default cache
backend. The difference between a \emph{fs} and \emph{cache} extension is that the
\emph{cache} extension may contain the \code{.}.
}
\examples{
  cache_ext()

}
\seealso{
\itemize{
\item \code{\link[=cache_backend]{cache_backend()}}
}
}
