% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_load}
\alias{cache_load}
\title{Load cached objects from cache directory}
\usage{
cache_load(objnames = "*", .cachedir = here(".cache-R"))
}
\arguments{
\item{objnames}{character list of object names to load.}

\item{.cachedir}{path to cache directory. Defaults to the directory named \code{.cache-R} at the project root.}
}
\description{
Load cached objects from cache directory
}
