% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ls.R
\name{cache_ls}
\alias{cache_ls}
\alias{cache_dir}
\alias{cache_info}
\title{cache_ls}
\usage{
cache_ls(cache = getOption("cache", "cache"), ...)

cache_dir(cache = getOption("cache", "cache"), ...)

cache_info(cache = getOption("cache", "cache"), ...)
}
\arguments{
\item{cache}{string; cache directory}

\item{...}{additional arguments passed to \code{\link[fs:dir_ls]{fs::dir_ls()}}}
}
\description{
List contents of the cache directory
}
\details{
Similar to \code{ls} but defaults to viewing the cache directory.
}
\seealso{
\itemize{
\item \code{\link[=cache]{cache()}}, \code{\link[=uncache]{uncache()}}
\item \link[fs:dir_ls() ]{fs::dir_ls() }
}
}
