% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_name.R
\name{cache_name}
\alias{cache_name}
\title{cache_name}
\usage{
cache_name(name = NULL)
}
\arguments{
\item{name}{string; name for the cache (Default: Null)}
}
\value{
string name for the cache directory.
}
\description{
Get/set the default \emph{name} for the cache directory (not the path). The
default is \code{cache}.
}
\details{
If used \code{name} is not provided or NULL, \code{cache_name} provides the default
cache name.

If \code{name} is provided, the default name is set to \code{name}.

The default cache directory name can be changes by setting
\code{options(cache=name)}.  The default is \code{cache} but alternative can be useful
such as \code{.cache} (hidden) or \code{rcache} (possibly avoiding conflicts).
}
\seealso{
\itemize{
\item \code{\link[=cache_path]{cache_path()}} which finds the path to the cache directory.
}
}
