% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_path}
\alias{cache_path}
\title{Get/set the cache path}
\usage{
cache_path(path)
}
\arguments{
\item{path}{path or string; location for the cache directory}
}
\value{
Always returns the value of \code{getOptions('cache.path')}
}
\description{
Get/set the cache path
}
\details{
Without a \code{path} argument or \code{path=NULL}, \code{cache_path()} returns the
(relative) path to the cache directory. This is provided by
\code{options(cache.path)}.

When \code{path} is given, the \code{cache_path} is set.  If \code{path} does not exist,
an error is thrown. The cache should be initialized with \code{cache_create(...)}
first. \code{NULL} is allowed and does not throw an error.
}
\seealso{
\itemize{
\item \code{\link[=cache_create]{cache_create()}}
\item \code{\link[fs:path_rel]{fs::path_rel()}}
}
}
