% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_path}
\alias{cache_path}
\title{cache_path}
\usage{
cache_path(path)
}
\arguments{
\item{path}{path or string; location for the cache directory}
}
\value{
\emph{Relative} path of the cache directory or \code{NULL} if doesn't exist.
}
\description{
Get/set the cache path
}
\details{
Without a \code{path} argument or \code{path=NULL}, \code{cache_path()} returns the
(relative) path to the cache directory. This is provided by
\code{options(cache.path)}.

With a \code{path} argument specified, \code{cache_path} set the path for the cache
directory. An error is raised if \code{path} does not exist. The cache should be
initialized with \code{cache_create(...)} first.
}
\seealso{
\itemize{
\item \code{\link[=cache_name]{cache_name()}}
\item \code{\link[=cache_find]{cache_find()}}
\item \code{\link[fs:path_rel]{fs::path_rel()}}
}
}
