% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_path}
\alias{cache_path}
\title{cache_path}
\usage{
cache_path(path = NULL)
}
\value{
\code{\link[fs:path_rel]{fs::path_rel()}} object of the \emph{relative} path to the cache directory.
Use \code{\link[fs:path_abs]{fs::path_abs()}} to get the absolute path.
}
\description{
Get/set the cache path
}
\details{
Without a \code{path} argument or \code{path=NULL}, \code{cache_path()} returns the
(relative) path to the cache directory. This is provided by
\code{options(cache.path)} or \code{\link[=cache_find]{cache_find()}}

With a \code{path} argument specified, \code{cache_path} set the path for the cache
directory.
}
\seealso{
\itemize{
\item \code{\link[=cache_name]{cache_name()}}
\item \code{\link[=cache_find]{cache_find()}}
}
}
