% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path_ext.R, R/cache_reader.R
\name{cache_path_ext}
\alias{cache_path_ext}
\title{Get the extension}
\usage{
cache_path_ext(path)

cache_path_ext(path)
}
\arguments{
\item{path}{string; path to file}
}
\description{
Get the \emph{registered} cache extension string from the path.
}
\details{
\code{\link[=cache_path_ext]{cache_path_ext()}} gets the entire extension from the cache path, either a
full path or simply a filename. The extension has to be known and registered
using \code{\link[=cache_add_ext]{cache_add_ext()}}.
}
\seealso{
\itemize{
\item \code{\link[=cache_add_ext]{cache_add_ext()}}
\item \code{\link[fs:path_ext]{fs::path_ext()}}
}
}
