% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_reader.R
\name{cache_reader}
\alias{cache_reader}
\alias{cache_writer}
\title{cache_reader, cache_writer}
\usage{
cache_reader()

cache_writer()
}
\description{
Get the read or write function for a path
}
\details{
\link{cache_reader{}} and \code{\link[=cache_writer]{cache_writer()}} return the function for reading
and writing based on a given path.
}
\seealso{
\itemize{
\item \code{\link[=cache_backend]{cache_backend()}}
}
}
