% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_rm.R
\name{cache_rm}
\alias{cache_rm}
\alias{cache_delete}
\title{cache_rm}
\usage{
cache_rm(..., .cache = cache_path())

cache_delete(..., .cache = cache_path())
}
\arguments{
\item{...}{single character vector of object names or comma separated list of
unquoted names.}

\item{.cache}{path; cache directory}
}
\description{
Remove item(s) from the cache
}
\details{
\code{\link[=cache_rm]{cache_rm()}} accepts either a character vector -or- a comma-separated list of
unquoted names.

\code{\link[=cache_delete]{cache_delete()}} is an alias for \code{\link[=cache_rm]{cache_rm()}}.
}
\examples{

\dontrun{
  cache(iris)
  cache_rm(iris)
  cache(iris)
  cache_rm('iris')

  cache(iris);cache(mtcars)
  cache_ls()
  cache_rm(iris,mtcars)
  cache_ls()


  cache(iris);cache(mtcars)
  cache_rm( c('iris','mtcars') )
  cache_ls()
}

}
