% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_rm.R
\name{cache_rm}
\alias{cache_rm}
\title{cache_rm}
\usage{
cache_rm(..., .cache = cache_find())
}
\arguments{
\item{...}{character; names of cached objects to permanently remove from
the cache}

\item{.cache}{path; cache directory}
}
\description{
Remove item(s) from the cache
}
