% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_use_rds.R
\name{cache_use_rds}
\alias{cache_use_rds}
\alias{cache_register_rds}
\title{Set the default cache method to rmat \code{RDS}.`}
\usage{
cache_use_rds()

cache_register_rds()
}
\description{
Cache objects using RDS \code{\link[base:saveRDS]{base::saveRDS()}}
}
\details{
This is the \emph{default} method for caching objects.

\code{cache_use_rds} sets RDS as the default caching mechanism to use RDS.
}
\examples{

\dontrun{
  cache_use_rds()
}

}
