% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_file.R, R/cached_name.R
\name{cached_file}
\alias{cached_file}
\alias{print.cached_file}
\alias{as_cached_file}
\alias{as_cached_file.default}
\alias{as_cached_file.character}
\alias{as_cached_file.fs_path}
\alias{as_cached_file.cached_name}
\alias{as_cached_name.default}
\title{Create a cached_file}
\usage{
cached_file(x)

\method{print}{cached_file}(x, ...)

as_cached_file(x, ext = cache_ext())

\method{as_cached_file}{default}(x, ext = cache_ext())

\method{as_cached_file}{character}(x, ext = cache_ext())

\method{as_cached_file}{fs_path}(x, ext = NULL)

\method{as_cached_file}{cached_name}(x, ext = cache_ext())

\method{as_cached_name}{default}(x)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments}

\item{ext}{ext or string used in the cache}
}
\value{
a \code{fs::fs_path} object of the filename. The extension
}
\description{
Create a name for an object.
}
\details{
A \code{cached_file} is a subclass of \code{fs_path} object without the directory.
It is the filename without the path information.
}
\examples{

 cached_file('iris.rds') # e.g. "iris.rds"
 as_cached_file('iris.rds')  # Warn: Don't include extension
 as_cached_file('iris', 'rds' )
 as_cached_file(iris)

  cached_name('iris') \%>\% as_cached_file()
}
