% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{conflicts}
\alias{conflicts}
\title{List cache conflicts}
\usage{
conflicts(name, ext = cache_ext())
}
\arguments{
\item{name}{quoted or unquoted name}
}
\value{
character vector of filenames -- conflicts
}
\description{
List filenames that conflict with name
}
\details{
\emph{cache} does not permit multiple files with the same name within a single
cache.  For example, you cannot have both \code{iris.rds} and \code{iris.csv}.
This enforces a best practice of uniquely naming your data set.

This uniqueness is enforced among files with registered backends.
Thus, if there is no backend for a \code{.txt} file,
it will not be identified as conflict.
}
\examples{

  conflicts(iris)  #
  conflicts('iris')

  has_conflict(iris)
  has_conflict('iris')

}
