% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{ext}
\alias{ext}
\title{filename/path extension}
\usage{
ext(path, exts = getOption("fs.exts"))
}
\arguments{
\item{path}{object such as character, vector, ; values to be converted to character. Character values
should omitting the leading \code{.} (period)}

\item{exts}{optional list of extensions, used when the extension contains a \code{.}}
}
\value{
a \code{\link[=fs_ext]{fs_ext()}} object.
}
\description{
Get the path or file extension allowing for extensions containing \code{.}
}
\details{
\code{\link[=ext]{ext()}} returns the extension of a file or path object. \code{ext} returns (the first of):
\itemize{
\item extension matching the \code{exts} arguments which uses the \code{fs.exts} argument by default.
\item character vector after final period.
}

The \code{\link[fs:path_ext]{fs::path_ext()}} takes only the considers the last.
}
\examples{

  files <- c("log.tar.gz","log.tar","log.tar.bz")
  known_exts=c("tar.gz","tar.bz")

  ext( files, exts=known_exts )
  ext( files )
  fs::path_ext( files )

}
\seealso{
\itemize{
\item \code{\link[fs:path_ext]{fs::path_ext()}} - This is nearly identical but does not allow extensions
containing \code{.} (periods).
}
}
