% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_root_safe.R, R/utils.R
\name{find_root_safe}
\alias{find_root_safe}
\alias{find_root_safe}
\title{Find Root or  NULL}
\usage{
find_root_safe(...)

find_root_safe(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[rprojroot:find_root]{rprojroot::find_root()}}}
}
\description{
Find the root or return NULL -- used when you don't know if you are in
a pack

Attempts to find the root returning NULL if it fails
}
\note{
this function is not exported
}
