% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_root_safe}
\alias{find_root_safe}
\title{Find the root, returning NULL if it is not safe}
\usage{
find_root_safe(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[rprojroot:find_root]{rprojroot::find_root()}}}
}
\description{
Attempts to find the root returning NULL if it fails
}
\note{
this function is not exported
}
