% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds.R
\name{save_rds}
\alias{save_rds}
\title{save_rds}
\usage{
save_rds(object, ...)
}
\arguments{
\item{object}{R object to serialize}

\item{...}{additional arguments passed to saveRDS}
}
\value{
\code{object}
}
\description{
A wrapper around \code{\link[base:saveRDS]{base::saveRDS()}} that returns the object invisibly for
use in pipes
}
\details{
Saves an object as an RDS file
}
\seealso{
\itemize{
\item \code{\link[base:saveRDS]{base::saveRDS()}}
\item readr::write_rds
}
}
